Fix virtualenv forking
`pypath-setup` is relaunching itself with the interpreter from the
virtualenv in use, but at this point it might be missing the
`pypath_setup` package that hosts its `main` function, and fails to
execute the subprocess of itself.

This commit removes the self-forking, which makes the script easier, and
less prone to such issues. We only run a command that fetches the
`sys.path` for us to build PYTHONPATH, and that's it.

Signed-off-by: Paul Maréchal <marechap.info@gmail.com>